Add REST Routes, RequestHandler and Serialize
The routes.php file is where we tell cakePHP that we want to map REST
routes for our notes as well as tell it to parse extensions (i.e. for
JSON or XML).  I left all on for now but you can limit it to just
specific ones.
RequestHandler is added to the Controller to tell it that we want too
look at the extension on the the request and produce json or xml as
such.