Add Security
Add Security to the User Model and Notes Model requiring
authentication.

Modify Notes Controller so a view request only shows the logged in
users notes.  The Notes Controller will only allow an edit or delete of
a note if it belongs to the logged in user.

Updated the User Add View so that a new user can’t set themselves as an
admin