Fixed Routes and Added Validation
Due to the nature of CakePHP being both HTML View based and REST based
the default controllers have some redirects for things like saving and
deleting.  This causes a conflict with the JSON responses.
I also added validation rules so that the title and description can not
be empty on the notes